fix: realtime convo and convo entires across spaces
